third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-feature-settings-computed-expected.txt

This is a testharness.js-based test.
[PASS] Property font-feature-settings value 'normal'
[PASS] Property font-feature-settings value '"dlig"'
[PASS] Property font-feature-settings value '"smcp"'
[PASS] Property font-feature-settings value '"c2sc"'
[PASS] Property font-feature-settings value '"liga" 0'
[FAIL] Property font-feature-settings value '"tnum", "hist"'
  assert_equals: expected "\"hist\", \"tnum\"" but got "\"tnum\", \"hist\""
[PASS] Property font-feature-settings value '"PKRN"'
[FAIL] Property font-feature-settings value '"dlig", "smcp", "dlig" 0'
  assert_equals: expected "\"dlig\" 0, \"smcp\"" but got "\"dlig\", \"smcp\", \"dlig\" 0"
Harness: the test ran to completion.
